+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 * Add Jsonnet support (#81)
 * Add Zig support (#82)
 * Add Erlang support (#83)
+* Add basic Pascal support (#84)
 
 ## 0.2.0
 > Released Sep 01, 2023